requirements.txt
openai>=0.27.0,<0.28.0
Pillow>=8.3.1,<8.4.0
requests>=2.26.0,<2.27.0

